+++ includes/cli/class-commands.php
@@ -33,7 +33,7 @@ class Commands extends WP_CLI_Command {
     public function status() {
         $roc = Plugin::instance();
 
-        require_once __DIR__ . '/../ui/diagnostics.php';
+        require_once __DIR__ . '/../diagnostics.php';
     }
 
     /**