+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Added `WP_Object_Cache::decr()` alias
+- Moved `diagnostics.php` file
 
 ## 2.0.20